requirements.txt
numpy
pandas
scipy
astropy
astroquery
sunpy==4.0.0
matplotlib
seaborn
streamlit==1.9.0
Pillow
stqdm
lxml
zeep
drms
tqdm
vedo
pytest-astropy
pytest-sugar
